+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
-
 ## [Unreleased]
+
+## [1.5.0]
 
 ### Fixed
 
@@ -50,7 +51,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 - Fix deprecation warning when running tests on PHP 8.1 and 8.2
 
-[Unreleased]: https://github.com/slunak/pushover-php/compare/v1.4.0...HEAD
+[Unreleased]: https://github.com/slunak/pushover-php/compare/v1.5.0...HEAD
+[1.5.0]: https://github.com/slunak/pushover-php/compare/v1.4.0...v1.5.0
 [1.4.0]: https://github.com/slunak/pushover-php/compare/v1.3.2...v1.4.0
 [1.3.2]: https://github.com/slunak/pushover-php/compare/v1.3.1...v1.3.2
 [1.3.1]: https://github.com/slunak/pushover-php/compare/v1.3.0...v1.3.1